test: - Move 'Rendering matching tasks' tests to QuickSearchTasksModal.test.ts
