+++ tests/Commands/QuickSearchTasks.test.ts
@@ -1,7 +1,6 @@
 import moment from 'moment';
 import { Notice } from 'obsidian';
 import {
-    QuickSearchTasksModal,
     findIncompleteTasksByDescription,
     findIncompleteTasksByDescriptionSubstring,
     openTaskAtSourceLocation,
@@ -12,7 +11,7 @@ import { getTaskLineAndFile } from '../../src/Obsidian/File';
 import { Status } from '../../src/Statuses/Status';
 import { TaskBuilder } from '../TestingTools/TaskBuilder';
 import { GlobalFilter } from '../../src/Config/GlobalFilter';
-import { fromLines, fromMarkdown } from '../TestingTools/TestHelpers';
+import { fromLines } from '../TestingTools/TestHelpers';
 import { GlobalQuery } from '../../src/Config/GlobalQuery';
 import type { PresetsMap } from '../../src/Query/Presets/Presets';
 import { resetSettings, updateSettings } from '../../src/Config/Settings';
@@ -361,72 +360,6 @@ describe('Finding matching tasks, sorting results in expected order', () => {
 
             expectSortsTasksInExpectedOrder(tasks, expectedOrder, (task: Task) => task.path);
         });
-    });
-});
-
-describe('Rendering matching tasks', () => {
-    function getCheckbox(element: HTMLDivElement): HTMLElement | null {
-        return element.querySelector('.tasks-quick-search-result-checkbox');
-    }
-
-    function getDescriptionTextContent(element: HTMLDivElement): string | null | undefined {
-        return element.querySelector('.tasks-quick-search-result-description')?.textContent;
-    }
-
-    function getLocationTextContent(element: HTMLDivElement): string | null | undefined {
-        return element.querySelector('.tasks-quick-search-result-location')?.textContent;
-    }
-
-    function getMetadata(element: HTMLDivElement): HTMLElement | null {
-        return element.querySelector('.tasks-quick-search-result-metadata');
-    }
-
-    it('should render the checkbox, description, location, and metadata elements', () => {
-        const modal = new QuickSearchTasksModal({} as any, () => tasks, jest.fn());
-        const element = document.createElement('div');
-
-        modal.renderSuggestion(writeReleaseNotes, element);
-
-        expect(getCheckbox(element)).not.toBeNull();
-        expect(getDescriptionTextContent(element)).toEqual('Write release notes');
-        expect(getLocationTextContent(element)).toEqual('Release.md · Preparation');
-        expect(getMetadata(element)).not.toBeNull();
-    });
-
-    it.each([
-        [false, '#task Do Stuff'],
-        [true, 'Do Stuff'],
-    ])(
-        'should honour "Remove global filter from description" setting: %s',
-        (removeGlobalFilter: boolean, expectedDescription: string) => {
-            GlobalFilter.getInstance().set('#task');
-            GlobalFilter.getInstance().setRemoveGlobalFilter(removeGlobalFilter);
-
-            const taskListWithGlobalFilter = fromMarkdown('- [ ] #task Do Stuff');
-
-            const modal = new QuickSearchTasksModal({} as any, () => taskListWithGlobalFilter, jest.fn());
-            const element = document.createElement('div');
-
-            modal.renderSuggestion(taskListWithGlobalFilter[0], element);
-
-            expect(getDescriptionTextContent(element)).toEqual(expectedDescription);
-        },
-    );
-
-    it('should only check the checkbox for completed tasks', () => {
-        const modal = new QuickSearchTasksModal({} as any, () => tasks, jest.fn());
-        const incompleteElement = document.createElement('div');
-        const completeElement = document.createElement('div');
-        const inProgressTask = new TaskBuilder().description('In progress task').status(Status.IN_PROGRESS).build();
-
-        modal.renderSuggestion(inProgressTask, incompleteElement);
-        modal.renderSuggestion(releaseCompleted, completeElement);
-
-        expect(incompleteElement.querySelector('input')).toMatchObject({ checked: false });
-        expect(incompleteElement.querySelector('input')?.getAttribute('aria-label')).toEqual(
-            'Task status: In Progress',
-        );
-        expect(completeElement.querySelector('input')).toMatchObject({ checked: true });
     });
 });
